requirements.txt
# Core dependencies
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
alembic>=1.12.0
hvac>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
python-multipart>=0.0.6
celery>=5.3.0
redis>=4.5.0
httpx>=0.24.0
paramiko>=3.1.0
aiohttp>=3.9.0
lxml>=4.9.0
PyYAML>=6.0
dnspython>=2.4.0
python-whois>=0.8.0
reportlab>=4.4.0
jinja2>=3.1.0

# Authentication and Security
pyotp>=2.9.0
qrcode>=7.4.2
pillow>=10.0.0

# Development
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0

# Optional - GVM/OpenVAS
python-gvm>=24.1.0
gvm-tools>=20.8.0
msgpack>=1.0.0
